+++ lib/packages/core/CHANGELOG.md
@@ -7,4 +7,5 @@ All notable changes to this project will be documented in this file.
 - Added `sendMoney` for generating the unsigned transaction, signing it, and broadcasting it.
 
 ### Account
+- Added `getAliases` to retrieve aliases for an account
 - Added `generateSendTransactionQRCode` and `generateSendTransactionQRCodeAddress` methods for generating a QR code image or URL, respectively.